Ensure all strings are marked with UTF-8 encoding.
From ruby 1.9.1 all strings must have an encoding, and since the strings
from sap already are UTF-8, change the string creation funtions to reflect
this. This change requires ruby 1.9.1 and newer.

Notice that while rb_enc_str_new is defined in ruby 1.9.1, the matching
rb_enc_str_new_cstr was not introduced before rupy 2.1.1, but since
only the SAP values may have UTF-8 characters in them, it is still ok to use
rb_str_new2 because the constants strings are ascii.

For strings that are originally fetched from ruby, store the corresponding
encoding, such that the string can be created with the correct encoding
when transferred back to the ruby runtime.
It may also be possible to just keep a reference to the original source string,
but it's unclear to me how that will interact with the garbage collectior.